refactor: create attributes for animation constants
There are two constants that have been made attributes in the
ProjectileMotionAnimation class.

One is ProjectileMotionAnimation.time_total_frames_multiplier, which is
used to determine how many frames an animation should take based on
on the computed total time of the ball's airtime irl..

Second is ProjectileMotionAnimation.upc, which is used to convert the
computed displacement into pixel values that will allow the ball cover
larger distances in the screen and make the animation look a bit better.