feat: add abort animation function
Now the animation is stopped and the ball's position reset when the
user resizes the window while the animation is going on.

Created a reset_ball_position() function to bring the ball back to
its default position.

Created a show_message() function to let the user know why the animation
ended. Could potentially be used for other things as well in the future.