feat: make 'abort animation' button functional
The 'Abort Animation' button now works when clicked and does abort
an animation while it is running. It brings the ball back to the initial
position and shows a small text in the upper left saying 'Manually
aborted the animation'.

The button is also disabled and enabled depending on whether there is an
animation running.