refactor: use str literals instead of tk constants
The constants NORMAL and DISABLED were being used when setting the state
of a button. Those have been replaced with "normal" and "disabled".

This is because the string literals were being used everywhere else
(when setting sticky), so I wanted to be consistent. I also saw this
StackOverflow answer by Brian Oakley, which gave me a good reason to do
so: https://stackoverflow.com/questions/48426912/importance-of-python-module-constants